Update of Example 1 using OM Notebook.
Unfortunately the scripting facility offered by OMNotebook is very limited, so I merged the 3 documents together to simplify the paths management.
Also I removed all "cd()" because the compilation generates too many files if put in the tutorial's folder. 
Also, as in previous tutorials, I assume that the tutorial package contains a "compiled" OpenIPSL.mo and Tutorial.mo in the archive.