Fixes #185 by removing the der(alpha) condition
In principle the boundaries where working but as it turns out the der(alpha) conditions were adding massive numerical chatter when alpha approached the limit. Removing the condition solves this whilst the limits still work. 
This can be tested by running `OpenIPSL.Examples.Branches.PSAT.PhaseShiftingTransformer_Test` and setting the limit to +- 50 °