Expanded instance name text rectangle
Expanded the text rectangles of groups 1 to 6, so that the group names are visible for a much larger part. 
Indeed  in case the instance name goes beyond the text rectangle both Dymola and OpenModelica use ellipsees, and show only part of the name.
Therefore it is recommended to use rectancles for instance names spanning the entire witdth of the icon. Previously the rectangles had much more limited horizontal widths.