Make line names more visible
The instance namee of lines are inside the line rectangles.
Prior to this commit, they were written in yellow with grey background. Unfortunately, in this way their visibility was not good. 
With this commit, the line names are now in black, and make the line icons much clearer.
Check the appedarence of IEEE_14_Buses for comparison.