Shorter and clearer generator names
Prior to this commit, the instancece names of the generators were very long, and therefore were truncated in diagrams. The length was not sue to special clarity purposes; it seemd just the consequence of some automatic naming algorithm.
This commit introduces shorter names, leaving for the generators names at least the same meaninfulness, but allowing them to be visible in diagrams without zooming.