Add n and p variant of the PwPin
Before this commit, pins at two ends of several components, such as transmission lines, had the same visual aspect. But they are different in meaning, since voltages and currents at them assume different values.
Therefore, much better is to visually differentiate the two pins, in compliance with what is done in the whole Modelica Standard Library (MSL).
With this commit, the positive pin is blue-filled while the negative is white filled, again in compliance with the general approach of MSL.